requirements.txt
fastapi
uvicorn
python-dotenv
SQLAlchemy>=2.0
asyncmy
pydantic
pandas
prisma
langchain
langchain_google_genai
pinecone
tqdm
langchain-google-vertexai
langchain-pinecone
google-generativeai